move entitlements into a core model file
